+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+## 3.2.0
 ## Added
 - default empty templates
 - ar-SA, eu-ES, fi-FI, he-IL, nb-NO, sl-SL and sr-Cyrl-RS empty file templates